fix: mini: Trim command spaces after TrimPrefix
In `func (v *groupView) commandParser`, the prefix of a command like "/debug store group_pk metadata" is removed with:

    strings.TrimPrefix(input, prefix+" ")

This is so that the `cmd` parameter for the command only has the args like "group_pk metadata" without the command prefix. But if the user simply enters "/debug store" then the call the `TrimPrefix` does not match. (There is no space after the prefix.) Therefore the `cmd` parameter for the command contains the prefix "/debug store". But `cmd` should be the empty string in this case. This commit changes to call `TrimPrefix` only with the prefix, and to trim the spaces after.

Signed-off-by: Jeff Thompson <jeff@thefirst.org>